refac: add style formatting for all files in src folder

+++ src/pages/Keys.vue
@@ -5,42 +5,30 @@
     :label="t('Search')"
     clearable
     hide-details
-    style="position: absolute; top: 2.5px;right: calc(25vw); width: 40vw; background: white;"
+    style="position: absolute; top: 2.5px; right: calc(25vw); width: 40vw; background: white"
   />
   <v-row>
     <v-col cols="auto">
       <h1>
         {{ t('APIKeys') }}
-        
+
         <v-btn
           v-has-perms="'write:keys'"
           prepend-icon="add"
           class="no-cap-btn bg-primary-600"
-          style="position: absolute; right: 10px;"
+          style="position: absolute; right: 10px"
           :text="t('AddApiKey')"
           @click="dialog = true"
         />
-        <key-add 
-          :dialog="dialog"
-          :item="selectedItem"
-          @close="close"
-        />
+        <key-add :dialog="dialog" :item="selectedItem" @close="close" />
       </h1>
     </v-col>
-    <v-col 
-      cols="auto"
-      align-self="center"
-    >
+    <v-col cols="auto" align-self="center">
       <keys-filter />
     </v-col>
   </v-row>
-  
 
-  <g-switch 
-    v-model="showExpired"
-    :label="t('ShowExpired')"
-    class="switch-primary"
-  />
+  <g-switch v-model="showExpired" :label="t('ShowExpired')" class="switch-primary" />
 
   <v-data-table
     v-model:sort-by="sortBy"
@@ -57,52 +45,25 @@
   >
     <template #[`item.key`]="{item}">
       {{ item.key }}
-      <v-btn
-        variant="text"
-        icon
-        @click.stop="() => copyKey(item)"
-      >
+      <v-btn variant="text" icon @click.stop="() => copyKey(item)">
         <v-icon>content_copy</v-icon>
-        <v-tooltip 
-          :text="keyTooltip"
-          activator="parent"
-          location="top"
-        />
+        <v-tooltip :text="keyTooltip" activator="parent" location="top" />
       </v-btn>
     </template>
     <template #[`item.status`]="{item}">
-      <v-icon :icon="item.status === 'active' ? 'check' : 'close' " />
+      <v-icon :icon="item.status === 'active' ? 'check' : 'close'" />
     </template>
-    
-    <template
-      v-for="desc in ['expireTime', 'lastUsedTime']"
-      #[`item.${desc}`]="{item}"
-      :key="desc"
-    >
-      <date-time
-        :value="item[desc as 'expireTime' | 'lastUsedTime']"
-      />
+
+    <template v-for="desc in ['expireTime', 'lastUsedTime']" #[`item.${desc}`]="{item}" :key="desc">
+      <date-time :value="item[desc as 'expireTime' | 'lastUsedTime']" />
     </template>
     <template #[`item.scopes`]="{item}">
-      <v-chip
-        v-for="scope in item.scopes"
-        :key="scope"
-        outline
-        size="small"
-        variant="flat"
-        class="chip"
-      >
+      <v-chip v-for="scope in item.scopes" :key="scope" outline size="small" variant="flat" class="chip">
         {{ scope }}
       </v-chip>
     </template>
     <template #[`item.actions`]="{item}">
-      <v-btn
-        v-has-perms.disable="'write:keys'"
-        icon="edit"
-        density="compact"
-        variant="text"
-        @click="editItem(item)"
-      />
+      <v-btn v-has-perms.disable="'write:keys'" icon="edit" density="compact" variant="text" @click="editItem(item)" />
       <v-btn
         v-has-perms.disable="'write:keys'"
         icon="content_copy"
@@ -122,21 +83,21 @@
 </template>
 
 <script lang="ts" setup>
-import type { Store } from '@/plugins/store/types';
-import type { SortBy } from '@/plugins/store/types/alerts-types';
-import type { Key } from '@/plugins/store/types/keys-types';
-import { computed, ref, watch } from 'vue';
-import { useI18n } from 'vue-i18n';
-import { useStore } from 'vuex';
+import type {Store} from '@/plugins/store/types'
+import type {SortBy} from '@/plugins/store/types/alerts-types'
+import type {Key} from '@/plugins/store/types/keys-types'
+import {computed, ref, watch} from 'vue'
+import {useI18n} from 'vue-i18n'
+import {useStore} from 'vuex'
 
 definePage({
   meta: {
-    title: "API Keys",
+    title: 'API Keys',
     requiresAuth: true
   }
-});
+})
 
-const { t } = useI18n()
+const {t} = useI18n()
 const store: Store = useStore()
 
 const dialog = ref(false)
@@ -146,20 +107,20 @@ const keyTooltip = ref(t('Copy'))
 const search = ref('')
 const sortBy = ref<SortBy[]>([{key: 'key', order: 'asc'}])
 
-const headers = ref<{title: string, key: keyof Key | 'actions', info?: string | string[], maxWidth?: number}[]>([
-  { title: t('APIKey'), key: 'key'},
-  { title: t('Active') + '?', key: 'status'},
-  { title: t('Scopes'), key: 'scopes'},
-  { title: t('Description'), key: 'text'},
-  { title: t('Expires'), key: 'expireTime'},
-  { title: t('Count'), key: 'count'},
-  { title: t('LastUsed'), key: 'lastUsedTime'},
-  { title: t('Actions'), key: 'actions' }
+const headers = ref<{title: string; key: keyof Key | 'actions'; info?: string | string[]; maxWidth?: number}[]>([
+  {title: t('APIKey'), key: 'key'},
+  {title: t('Active') + '?', key: 'status'},
+  {title: t('Scopes'), key: 'scopes'},
+  {title: t('Description'), key: 'text'},
+  {title: t('Expires'), key: 'expireTime'},
+  {title: t('Count'), key: 'count'},
+  {title: t('LastUsed'), key: 'lastUsedTime'},
+  {title: t('Actions'), key: 'actions'}
 ])
-const items = computed(() => store.state.keys.items.filter((h) => showExpired.value || h.status ==='active'))
+const items = computed(() => store.state.keys.items.filter(h => showExpired.value || h.status === 'active'))
 const refresh = computed(() => store.state.refresh)
 
-watch(refresh, (val) => {
+watch(refresh, val => {
   if (!val) return
   getItems()
 })
@@ -176,24 +137,21 @@ async function copyKey(item: Key) {
     console.error('Failed to copy key: ', error)
     keyTooltip.value = t('CopyFail')
   }
-  setTimeout(() => keyTooltip.value = t('Copy'), 2000)
+  setTimeout(() => (keyTooltip.value = t('Copy')), 2000)
 }
 
-
 function deleteItem(item: Key) {
-  confirm(t('ConfirmDelete')) &&
-  store.dispatch('keys/deleteKey', item.id!)
+  confirm(t('ConfirmDelete')) && store.dispatch('keys/deleteKey', item.id!)
   getItems()
 }
 
 function editItem(item: Key) {
-  selectedItem.value =  item
+  selectedItem.value = item
   dialog.value = true
 }
 
-
 function copyItem(item: Key) {
-  selectedItem.value =  {...item, id: undefined, key: undefined}
+  selectedItem.value = {...item, id: undefined, key: undefined}
   dialog.value = true
 }
 
@@ -201,7 +159,6 @@ function close() {
   dialog.value = false
   selectedItem.value = null
 }
-
 
 getItems()
 </script>
